+++ src/main/java/com/clap/pause/dto/post/request/ImageVoteRequest.java
@@ -1,5 +1,6 @@
-package com.clap.pause.dto.post;
+package com.clap.pause.dto.post.request;
 
+import com.clap.pause.dto.post.request.ImageVoteOptionRequest;
 import com.clap.pause.model.PostCategory;
 import com.clap.pause.model.PostType;
 import jakarta.validation.constraints.NotBlank;
@@ -8,7 +9,7 @@ import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 
-public record TextVoteRequest(
+public record ImageVoteRequest(
         @NotNull(message = "학과 그룹은 반드시 선택되어야 합니다.")
         Long departmentGroupId,
         @NotBlank(message = "제목은 최소 1자 이상이어야 합니다.")
@@ -19,7 +20,7 @@ public record TextVoteRequest(
         PostCategory postCategory,
         @NotNull(message = "글 타입은 반드시 선택되어야 합니다.")
         PostType postType,
-        @NotEmpty(message = "텍스트 투표 항목은 최소 하나 이상 존재해야 합니다.")
-        List<TextVoteOptionRequest> options
+        @NotEmpty(message = "이미지 투표 항목은 최소 하나 이상 존재해야 합니다.")
+        List<ImageVoteOptionRequest> options
 ) {
 }
