+++ src/main/java/com/clap/pause/repository/post/PostRepository.java
@@ -1,9 +1,9 @@
-package com.clap.pause.repository;
+package com.clap.pause.repository.post;
 
 import com.clap.pause.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
 }
